Virtual Machine lookup methods
Added methods to search virtual machines in all VDCs and VAPPs.
Since this feature is not directly implemented in the Abiquo API,
the client will get all vms of all vapps in all vdcs. All these
requests will be executed concurrently, to improve performance,
and the results aggregated and filtered to return the requested
vms.