Machine lookup methods
Added methods to search physical machines in all DCs and racks.
Since this feature is not directly implemented in the Abiquo API,
the client will get all racks in all datacenters. All these
requests will be executed concurrently, to improve performance,
and the results aggregated and filtered to return the requested
machines.