Upgraded to jclouds 1.5.0-SNAPSHOT
The EventBusModule has been removed since a pull request has been made
to include it as part of the jclouds-core project. 1.5.0-alpha.3 had a
minimal EventBus configuration that conflicted with the one in Abiquo,
so the pull request was made to replace it with the Abiquo one.

Also changes were made in jclouds-core to make the JAXBParser
pretty-printing configurable via context properties.

These changes exist in the abiquo branch of the nacx/jclouds fork until
the pull requests are merged in the jclouds master branch.