Empty server response. Fixed.
Creation of Private Networks raised 'NullPointerException'. Public and
external networks didn't handle well the API error and the message
was empty.

@nacx, I've changed this entity because the vlan tag can be null in
private networks at creation-time , and the signature of the method
accepted a builtin 'int' instead of an Integer.

This bug was introduced when I migrate the method of creating VLANs
from the flex-client-stub from old Wink Rest Client to JClouds provider.

cc / Jira - [ABICLOUDPREMIUM-3960]