ABIQUOJC-4 Fixed issues related to vm names
Abiquo automatically generates a name for the virtual machines, to avoid
name collision in the hypervisor. However, users can set a name too to
their virtual machines to easily identify them.

The "name" field in the VirtualMachine domain object has been
substituted by the "nameLabel" and the "internalName" fields. The first
one is the one that can be manipulated by the user, and the second one
will be assigned by abiquo when the VM is deployed.

Also changed all methods that retrieve a VMs to use the
VirtualMachineWithNodeExtendedDto instead of the VirtualMachineDto to
make sure that all necessary information is always retrieved.