Refactored query options classes.
Refactored all query options classes to subclass BaseHttpRequestOptions.
This is the default class in jclouds to build the requests parameters
and using it makes the AppendOptionsToPath binder redundant.

Deleted the QueryOptions abstract class and moved all filtering
attributes to the FilterOptions class.

Deleted also the AppendOptionsToPathBinder and updated all AsyncApi
interfaces accordingly.